+++ aimiaobi-20230801/alibabacloud_aimiaobi20230801/models.py
@@ -944,6 +944,140 @@ class AddAuditTermsResponse(TeaModel):
         return self
 
 
+class AddDatasetDocumentRequestDocumentMetadataAsrSentences(TeaModel):
+    def __init__(
+        self,
+        end_time: int = None,
+        start_time: int = None,
+        text: str = None,
+    ):
+        self.end_time = end_time
+        self.start_time = start_time
+        self.text = text
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.end_time is not None:
+            result['EndTime'] = self.end_time
+        if self.start_time is not None:
+            result['StartTime'] = self.start_time
+        if self.text is not None:
+            result['Text'] = self.text
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('EndTime') is not None:
+            self.end_time = m.get('EndTime')
+        if m.get('StartTime') is not None:
+            self.start_time = m.get('StartTime')
+        if m.get('Text') is not None:
+            self.text = m.get('Text')
+        return self
+
+
+class AddDatasetDocumentRequestDocumentMetadataVideoShots(TeaModel):
+    def __init__(
+        self,
+        end_time: int = None,
+        start_time: int = None,
+        text: str = None,
+    ):
+        self.end_time = end_time
+        self.start_time = start_time
+        self.text = text
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.end_time is not None:
+            result['EndTime'] = self.end_time
+        if self.start_time is not None:
+            result['StartTime'] = self.start_time
+        if self.text is not None:
+            result['Text'] = self.text
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('EndTime') is not None:
+            self.end_time = m.get('EndTime')
+        if m.get('StartTime') is not None:
+            self.start_time = m.get('StartTime')
+        if m.get('Text') is not None:
+            self.text = m.get('Text')
+        return self
+
+
+class AddDatasetDocumentRequestDocumentMetadata(TeaModel):
+    def __init__(
+        self,
+        asr_sentences: List[AddDatasetDocumentRequestDocumentMetadataAsrSentences] = None,
+        text: str = None,
+        video_shots: List[AddDatasetDocumentRequestDocumentMetadataVideoShots] = None,
+    ):
+        self.asr_sentences = asr_sentences
+        self.text = text
+        self.video_shots = video_shots
+
+    def validate(self):
+        if self.asr_sentences:
+            for k in self.asr_sentences:
+                if k:
+                    k.validate()
+        if self.video_shots:
+            for k in self.video_shots:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['AsrSentences'] = []
+        if self.asr_sentences is not None:
+            for k in self.asr_sentences:
+                result['AsrSentences'].append(k.to_map() if k else None)
+        if self.text is not None:
+            result['Text'] = self.text
+        result['VideoShots'] = []
+        if self.video_shots is not None:
+            for k in self.video_shots:
+                result['VideoShots'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.asr_sentences = []
+        if m.get('AsrSentences') is not None:
+            for k in m.get('AsrSentences'):
+                temp_model = AddDatasetDocumentRequestDocumentMetadataAsrSentences()
+                self.asr_sentences.append(temp_model.from_map(k))
+        if m.get('Text') is not None:
+            self.text = m.get('Text')
+        self.video_shots = []
+        if m.get('VideoShots') is not None:
+            for k in m.get('VideoShots'):
+                temp_model = AddDatasetDocumentRequestDocumentMetadataVideoShots()
+                self.video_shots.append(temp_model.from_map(k))
+        return self
+
+
 class AddDatasetDocumentRequestDocumentMultimodalMedias(TeaModel):
     def __init__(
         self,
@@ -994,6 +1128,7 @@ class AddDatasetDocumentRequestDocument(TeaModel):
         extend_1: str = None,
         extend_2: str = None,
         extend_3: str = None,
+        metadata: AddDatasetDocumentRequestDocumentMetadata = None,
         multimodal_index_name: str = None,
         multimodal_medias: List[AddDatasetDocumentRequestDocumentMultimodalMedias] = None,
         pub_time: str = None,
@@ -1010,6 +1145,7 @@ class AddDatasetDocumentRequestDocument(TeaModel):
         self.extend_1 = extend_1
         self.extend_2 = extend_2
         self.extend_3 = extend_3
+        self.metadata = metadata
         self.multimodal_index_name = multimodal_index_name
         self.multimodal_medias = multimodal_medias
         self.pub_time = pub_time
@@ -1019,6 +1155,8 @@ class AddDatasetDocumentRequestDocument(TeaModel):
         self.url = url
 
     def validate(self):
+        if self.metadata:
+            self.metadata.validate()
         if self.multimodal_medias:
             for k in self.multimodal_medias:
                 if k:
@@ -1046,6 +1184,8 @@ class AddDatasetDocumentRequestDocument(TeaModel):
             result['Extend2'] = self.extend_2
         if self.extend_3 is not None:
             result['Extend3'] = self.extend_3
+        if self.metadata is not None:
+            result['Metadata'] = self.metadata.to_map()
         if self.multimodal_index_name is not None:
             result['MultimodalIndexName'] = self.multimodal_index_name
         result['MultimodalMedias'] = []
@@ -1082,6 +1222,9 @@ class AddDatasetDocumentRequestDocument(TeaModel):
             self.extend_2 = m.get('Extend2')
         if m.get('Extend3') is not None:
             self.extend_3 = m.get('Extend3')
+        if m.get('Metadata') is not None:
+            temp_model = AddDatasetDocumentRequestDocumentMetadata()
+            self.metadata = temp_model.from_map(m['Metadata'])
         if m.get('MultimodalIndexName') is not None:
             self.multimodal_index_name = m.get('MultimodalIndexName')
         self.multimodal_medias = []
@@ -15058,12 +15201,14 @@ class GetDatasetDocumentRequest(TeaModel):
         dataset_name: str = None,
         doc_id: str = None,
         doc_uuid: str = None,
+        include_fields: List[str] = None,
         workspace_id: str = None,
     ):
         self.dataset_id = dataset_id
         self.dataset_name = dataset_name
         self.doc_id = doc_id
         self.doc_uuid = doc_uuid
+        self.include_fields = include_fields
         # This parameter is required.
         self.workspace_id = workspace_id
 
@@ -15084,6 +15229,8 @@ class GetDatasetDocumentRequest(TeaModel):
             result['DocId'] = self.doc_id
         if self.doc_uuid is not None:
             result['DocUuid'] = self.doc_uuid
+        if self.include_fields is not None:
+            result['IncludeFields'] = self.include_fields
         if self.workspace_id is not None:
             result['WorkspaceId'] = self.workspace_id
         return result
@@ -15098,8 +15245,202 @@ class GetDatasetDocumentRequest(TeaModel):
             self.doc_id = m.get('DocId')
         if m.get('DocUuid') is not None:
             self.doc_uuid = m.get('DocUuid')
+        if m.get('IncludeFields') is not None:
+            self.include_fields = m.get('IncludeFields')
         if m.get('WorkspaceId') is not None:
             self.workspace_id = m.get('WorkspaceId')
+        return self
+
+
+class GetDatasetDocumentShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        dataset_id: int = None,
+        dataset_name: str = None,
+        doc_id: str = None,
+        doc_uuid: str = None,
+        include_fields_shrink: str = None,
+        workspace_id: str = None,
+    ):
+        self.dataset_id = dataset_id
+        self.dataset_name = dataset_name
+        self.doc_id = doc_id
+        self.doc_uuid = doc_uuid
+        self.include_fields_shrink = include_fields_shrink
+        # This parameter is required.
+        self.workspace_id = workspace_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.dataset_id is not None:
+            result['DatasetId'] = self.dataset_id
+        if self.dataset_name is not None:
+            result['DatasetName'] = self.dataset_name
+        if self.doc_id is not None:
+            result['DocId'] = self.doc_id
+        if self.doc_uuid is not None:
+            result['DocUuid'] = self.doc_uuid
+        if self.include_fields_shrink is not None:
+            result['IncludeFields'] = self.include_fields_shrink
+        if self.workspace_id is not None:
+            result['WorkspaceId'] = self.workspace_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DatasetId') is not None:
+            self.dataset_id = m.get('DatasetId')
+        if m.get('DatasetName') is not None:
+            self.dataset_name = m.get('DatasetName')
+        if m.get('DocId') is not None:
+            self.doc_id = m.get('DocId')
+        if m.get('DocUuid') is not None:
+            self.doc_uuid = m.get('DocUuid')
+        if m.get('IncludeFields') is not None:
+            self.include_fields_shrink = m.get('IncludeFields')
+        if m.get('WorkspaceId') is not None:
+            self.workspace_id = m.get('WorkspaceId')
+        return self
+
+
+class GetDatasetDocumentResponseBodyDataMetadataAsrSentences(TeaModel):
+    def __init__(
+        self,
+        end_time: int = None,
+        start_time: int = None,
+        text: str = None,
+    ):
+        self.end_time = end_time
+        self.start_time = start_time
+        self.text = text
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.end_time is not None:
+            result['EndTime'] = self.end_time
+        if self.start_time is not None:
+            result['StartTime'] = self.start_time
+        if self.text is not None:
+            result['Text'] = self.text
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('EndTime') is not None:
+            self.end_time = m.get('EndTime')
+        if m.get('StartTime') is not None:
+            self.start_time = m.get('StartTime')
+        if m.get('Text') is not None:
+            self.text = m.get('Text')
+        return self
+
+
+class GetDatasetDocumentResponseBodyDataMetadataVideoShots(TeaModel):
+    def __init__(
+        self,
+        end_time: int = None,
+        start_time: int = None,
+        text: str = None,
+    ):
+        self.end_time = end_time
+        self.start_time = start_time
+        self.text = text
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.end_time is not None:
+            result['EndTime'] = self.end_time
+        if self.start_time is not None:
+            result['StartTime'] = self.start_time
+        if self.text is not None:
+            result['Text'] = self.text
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('EndTime') is not None:
+            self.end_time = m.get('EndTime')
+        if m.get('StartTime') is not None:
+            self.start_time = m.get('StartTime')
+        if m.get('Text') is not None:
+            self.text = m.get('Text')
+        return self
+
+
+class GetDatasetDocumentResponseBodyDataMetadata(TeaModel):
+    def __init__(
+        self,
+        asr_sentences: List[GetDatasetDocumentResponseBodyDataMetadataAsrSentences] = None,
+        text: str = None,
+        video_shots: List[GetDatasetDocumentResponseBodyDataMetadataVideoShots] = None,
+    ):
+        self.asr_sentences = asr_sentences
+        self.text = text
+        self.video_shots = video_shots
+
+    def validate(self):
+        if self.asr_sentences:
+            for k in self.asr_sentences:
+                if k:
+                    k.validate()
+        if self.video_shots:
+            for k in self.video_shots:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['AsrSentences'] = []
+        if self.asr_sentences is not None:
+            for k in self.asr_sentences:
+                result['AsrSentences'].append(k.to_map() if k else None)
+        if self.text is not None:
+            result['Text'] = self.text
+        result['VideoShots'] = []
+        if self.video_shots is not None:
+            for k in self.video_shots:
+                result['VideoShots'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.asr_sentences = []
+        if m.get('AsrSentences') is not None:
+            for k in m.get('AsrSentences'):
+                temp_model = GetDatasetDocumentResponseBodyDataMetadataAsrSentences()
+                self.asr_sentences.append(temp_model.from_map(k))
+        if m.get('Text') is not None:
+            self.text = m.get('Text')
+        self.video_shots = []
+        if m.get('VideoShots') is not None:
+            for k in m.get('VideoShots'):
+                temp_model = GetDatasetDocumentResponseBodyDataMetadataVideoShots()
+                self.video_shots.append(temp_model.from_map(k))
         return self
 
 
@@ -15109,9 +15450,12 @@ class GetDatasetDocumentResponseBodyData(TeaModel):
         content: str = None,
         disable_handle_multimodal_media: bool = None,
         doc_id: str = None,
+        doc_type: str = None,
         doc_uuid: str = None,
+        metadata: GetDatasetDocumentResponseBodyDataMetadata = None,
         pub_time: str = None,
         source_from: str = None,
+        status: int = None,
         summary: str = None,
         title: str = None,
         url: str = None,
@@ -15119,15 +15463,19 @@ class GetDatasetDocumentResponseBodyData(TeaModel):
         self.content = content
         self.disable_handle_multimodal_media = disable_handle_multimodal_media
         self.doc_id = doc_id
+        self.doc_type = doc_type
         self.doc_uuid = doc_uuid
+        self.metadata = metadata
         self.pub_time = pub_time
         self.source_from = source_from
+        self.status = status
         self.summary = summary
         self.title = title
         self.url = url
 
     def validate(self):
-        pass
+        if self.metadata:
+            self.metadata.validate()
 
     def to_map(self):
         _map = super().to_map()
@@ -15141,12 +15489,18 @@ class GetDatasetDocumentResponseBodyData(TeaModel):
             result['DisableHandleMultimodalMedia'] = self.disable_handle_multimodal_media
         if self.doc_id is not None:
             result['DocId'] = self.doc_id
+        if self.doc_type is not None:
+            result['DocType'] = self.doc_type
         if self.doc_uuid is not None:
             result['DocUuid'] = self.doc_uuid
+        if self.metadata is not None:
+            result['Metadata'] = self.metadata.to_map()
         if self.pub_time is not None:
             result['PubTime'] = self.pub_time
         if self.source_from is not None:
             result['SourceFrom'] = self.source_from
+        if self.status is not None:
+            result['Status'] = self.status
         if self.summary is not None:
             result['Summary'] = self.summary
         if self.title is not None:
@@ -15163,12 +15517,19 @@ class GetDatasetDocumentResponseBodyData(TeaModel):
             self.disable_handle_multimodal_media = m.get('DisableHandleMultimodalMedia')
         if m.get('DocId') is not None:
             self.doc_id = m.get('DocId')
+        if m.get('DocType') is not None:
+            self.doc_type = m.get('DocType')
         if m.get('DocUuid') is not None:
             self.doc_uuid = m.get('DocUuid')
+        if m.get('Metadata') is not None:
+            temp_model = GetDatasetDocumentResponseBodyDataMetadata()
+            self.metadata = temp_model.from_map(m['Metadata'])
         if m.get('PubTime') is not None:
             self.pub_time = m.get('PubTime')
         if m.get('SourceFrom') is not None:
             self.source_from = m.get('SourceFrom')
+        if m.get('Status') is not None:
+            self.status = m.get('Status')
         if m.get('Summary') is not None:
             self.summary = m.get('Summary')
         if m.get('Title') is not None: